ingest/pyproject.toml
[project]
name = "ingest"
version = "0.1.0"
description = "Gmail → Claude → Vikunja task ingest service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "google-cloud-pubsub>=2.26",
    "google-auth>=2.35",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.150",
    "anthropic>=0.39",
    "httpx>=0.27",
    "psycopg[binary,pool]>=3.2",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "pytest-asyncio>=0.24", "ruff>=0.7", "mypy>=1.13"]

[project.scripts]
ingest = "ingest.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ingest"]

[tool.ruff]
line-length = 100
target-version = "py311"
